+++ changelog.md
@@ -1,2 +1,5 @@
+# v1.0.1
+    - Fixed issue with LevelInfoLayer crashing if its your own level and you have Copy Hack enabled.
+    - Fixed Force Platformer Mode and Gravity not working on Windows
 # v1.0.0
     - Geode public release